+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 * Not yet released.
 * New dependency on weblate-language-data module.
 * Reduced amount of false positives on language codes inside a filename.
+* Improved iOS strings encoding detection.
 
 2.2
 ---